commit feature for currencics

+++ currencies/src/models.rs
@@ -2,7 +2,7 @@ use super::schema::currency_store;
 use chrono::NaiveDateTime;
 use diesel::expression::*;
 
-#[derive(Insertable)]
+#[derive(Insertable, Debug)]
 #[table_name = "currency_store"]
 pub struct NewCurrencyStore<'a> {
     pub id: &'a str,
@@ -10,7 +10,7 @@ pub struct NewCurrencyStore<'a> {
     pub txid: &'a str,
     pub update_time: &'a NaiveDateTime,
     pub last_owner_id: &'a str,
-    pub status: i32,
+    pub status: i16,
 }
 
 #[derive(Queryable, Debug, Clone)]
@@ -20,5 +20,5 @@ pub struct CurrencyStore {
     pub txid: String,
     pub update_time: NaiveDateTime,
     pub last_owner_id: String,
-    pub status: i32,
+    pub status: i16,
 }
